pyproject.toml
[project]
name = "simplebank"
version = "0.1.0"
description = "SimpleBank REST API"
authors = [
    {name = "Developer",email = "dev@simplebank.com"}
]
requires-python = "^3.14"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
